# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UAT Test Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /test-plans:
    get:
      operationId: listTestPlans
      tags: [testPlans]
      summary: List test plans
      parameters:
        - name: testCycleId
          in: query
          required: false
          schema:
            type: integer
        - name: releaseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of test plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestPlan"
    post:
      operationId: createTestPlan
      tags: [testPlans]
      summary: Create a test plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestPlanBody"
      responses:
        "201":
          description: Created test plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestPlan"
  /test-plans/{id}:
    get:
      operationId: getTestPlan
      tags: [testPlans]
      summary: Get test plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test plan details with scenarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestPlanDetail"
    patch:
      operationId: updateTestPlan
      tags: [testPlans]
      summary: Update test plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestPlanBody"
      responses:
        "200":
          description: Updated test plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestPlan"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, department, isActive, createdAt]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
      required: [name, email, role, department]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        isActive:
          type: boolean
...
    TestPlan:
      type: object
      properties:
        id:
          type: integer
        testCycleId:
          type: integer
        name:
          type: string
        description:
          type: string
        module:
          type: string
        status:
          type: string
        entryCriteria:
          type: string
        exitCriteria:
          type: string
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, testCycleId, name, description, module, status, createdAt]
    TestPlanDetail:
      type: object
      properties:
        id:
          type: integer
        testCycleId:
          type: integer
        name:
          type: string
        description:
          type: string
        module:
          type: string
        status:
          type: string
        entryCriteria:
          type: string
        exitCriteria:
          type: string
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        scenarios:
          type: array
          items:
            $ref: "#/components/schemas/TestScenario"
        stats:
          $ref: "#/components/schemas/PlanStats"
      required: [id, testCycleId, name, description, module, status, createdAt, scenarios, stats]
    PlanStats:
      type: object
      properties:
        totalScenarios:
          type: integer
        totalCases:
          type: integer
        passed:
          type: integer
        failed:
          type: integer
        blocked:
          type: integer
        notStarted:
          type: integer
        passRate:
          type: number
      required: [totalScenarios, totalCases, passed, failed, blocked, notStarted, passRate]
    CreateTestPlanBody:
      type: object
      properties:
        testCycleId:
          type: integer
        name:
          type: string
        description:
          type: string
        module:
          type: string
        status:
          type: string
        entryCriteria:
          type: string
        exitCriteria:
          type: string
        ownerId:
          type: ["integer", "null"]
      required: [testCycleId, name, description, module]
    UpdateTestPlanBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        module:
          type: string
        status:
          type: string
        entryCriteria:
          type: string
        exitCriteria:
          type: string
        ownerId:
          type: ["integer", "null"]
    TestScenario:
      type: object
      properties:
        id:
          type: integer
        testPlanId:
          type: integer
        title:
          type: string
        description:
          type: string
        businessProcess:
          type: string
        priority:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, testPlanId, title, description, businessProcess, priority, createdAt]
...