# lib/api-zod/src/generated/types/updateTestPlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UAT Test Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTestPlanBody {
  name?: string;
  description?: string;
  module?: string;
  status?: string;
  entryCriteria?: string;
  exitCriteria?: string;
  /** @nullable */
  ownerId?: number | null;
}